Issue #235: Make master factory injectable only on test WebFront
This is a preperatory refactoring. Because the plan is to set the
context on to the frontend factory during bootstrap, the frontend
factory always needs to be instantiaded when the web front is
created. To document this, the master factory is no longer injectable
since that was the switch if the frontend factory should be created
and registered or not.