Issue #235: Set Context instance on frontend factory
In order to be able to remove the request handler builders, this
change adds the ability to set the context on the frontend factory.
The request handlers now will be instantiable from the frontend
factory directly.
The WebFront is responsible for setting the context on the
frontend factory during bootstrap.